Don't modify GCC diagnostics when using Clang
The -Wclobbered warnings were fixed in #29642 by disabling that
particular GCC diagnostic for particular blocks of code. However, it was
guarded by a check on __GNUC__, which Clang also defines, and Clang
doesn't understand that diagnostic group, so it emits its own warning.
This changes the guards to be GCC-only by checking for __GNUC__ and not
__clang__.